crypto/Crypto Masquerade/requirements.txt
cryptography==39.0.0
pycryptodome==3.15.0